+++ src/core/injections/results_based/techniques/classic/cb_injector.py
@@ -24,7 +24,6 @@ from src.thirdparty.six.moves import html_parser as _html_parser
 from src.utils import menu
 from src.utils import settings
 from src.thirdparty.colorama import Fore, Back, Style, init
-from src.core.requests import tor
 from src.core.requests import proxy
 from src.core.requests import headers
 from src.core.requests import requests