+++ src/core/injections/semiblind/techniques/tempfile_based/tfb_injector.py
@@ -25,7 +25,6 @@ from src.utils import menu
 from src.utils import settings
 from src.utils import common
 from src.thirdparty.colorama import Fore, Back, Style, init
-from src.core.requests import tor
 from src.core.requests import proxy
 from src.core.requests import headers
 from src.core.requests import requests